Sentinel1

Main.Sentinel1.AzimuthFmRate — Type

AzimuthFmRate

returns structure of AzimuthFmRate from metadata in .xml AzimuthFmRate is calculated for each burst, and is therefore saved in each burst

Main.Sentinel1.AzimuthFmRate — Method

" AzimuthFmRate

Constucture for the AzimuthFmRate structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the AzimuthFmRate data for a single burst as a structure. 
The AzimuthFmRate structure returns the following:
        burstNumber::Int64
        fmTimesDiff::Vector{Millisecond}
        bestFmIndex::Int64
        azimuthFmRatePolynomial::Vector{Float64}
        azimuthFmRateT0::Float64

The burstNumber can be used as a key to other burst specific strucutres, e.g., the Burst data.

Input:
    metaDict[dict]: a dictionary of the metadata.
    dopplerCentroid[DopplerCentroid]: DopplerCentroid Structure

output:
    azimuthFmRate[structure of AzimuthFmRate]

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., azimuthFmRatePolynomial is redudant in the later processing. Could then be deleted.

Main.Sentinel1.Burst — Type

Burst

returns structure of Burst from metadata in .xml Burst contain information from DopplerCentroid and AzimuthFmRate

Main.Sentinel1.Burst — Type

" Burst

Constucture for the Burst structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Burst specific data for a single burst as a structure. 
The Burst structure returns the following:
    burstNumber::Int64
    azimuthTime::DateTime : Zero Doppler azimuth time of the first line of this burst [UTC]. 
    sensingTime::DateTime : Sensing time of the first input line of this burst [UTC].
    azimuthAnxTime::Float64 : Zero Doppler azimuth time of the first line of this burst relative to the Ascending Node Crossing (ANX) time. [milli sec].
    byteOffset::Int64.:  Byte offset of this burst within the image MDS.
    firstValidSample::Vector{Int64}: An array of integers indicating the offset of the first valid image sample within each range line. This array contains count attribute integers, equal to the linesPerBurst field (i.e. one value per range line within the burst), separated by spaces. If a range line does not contain any valid image samples, the integer is set to -1.
    lastValidSample::Vector{Int64}: An array of integers indicating the offset of the last valid image sample within each range line. This array contains count attribute integers, equal to the linesPerBurst (i.e. one value per range line within the burst), separated by spaces. If a range line does not contain any valid image samples, the integer is set to -1.
    burstId::Int64
    absoluteBurstId::Int64
    fmTimesDiff::Vector{Millisecond}
    bestFmIndex::Int64
    azimuthFmRatePolynomial::Vector{Float64}
    azimuthFmRateT0::Float64
    numberOfDopplerCentroids::Int64
    burstTime::Millisecond
    burstMidTime::Millisecond
    dcTimeDifferences::Vector{Millisecond}
    bestDcIndex::Int64
    dataDcPolynomial::Vector{Float64}
    dcT0::Float64
    firstLineMosaic::Int64

Input:
    metaDict[dict]: a dictionary of the metadata.
    burstNumber[Int]: Integer value of burst number.

output:
    Burst[structure of Burst]

Main.Sentinel1.BurstsInfo — Type

BurstsInfo

Returns a structure, BurstsInfo, containing info of each burst. For each bust, the following is saved in a Vector: - Burst: Structure of Burst.

Main.Sentinel1.BurstsInfo — Method

" BurstsInfo

Constucture for the BurstsInfo structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the BurstsInfo data for a all bursts as a structure. 
The burst info has the info and derived data for each bursts.

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
burstinfo[structure of BurstsInfo]

Note:
    Instead of vector{strucutre} consider another type???

Main.Sentinel1.DopplerCentroid — Type

DopplerCentroid

returns structure of DopplerCentroid from metadata in .xml DopplerCentroid is calculated for each burst, and is therefore saved in each burst

Main.Sentinel1.DopplerCentroid — Method

" DopplerCentroid

Constucture for the DopplerCentroid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the DopplerCentroid data for a single burst as a structure. 
The DopplerCentroid structure returns the following:
    numberOfDopplerCentroids::Int64
    burstTime::Millisecond
    burstMidTime::Millisecond
    dcTimeDifferences::Vector{Millisecond}
    bestDcIndex::Int64
    dataDcPolynomial::Vector{Float64}
    dcT0::Float64
    firstLineMosaic::Int64

Input:
    metaDict[dict]: a dictionary of the metadata.
    burst.azimuthTime
    header.startTime
    swathtiming.linesPerBurst
    imageInformation.azimuthFrequency

output:
    DopplerCentroid[structure of DopplerCentroid]

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., dataDcPolynomial is redudant in the later processing. Could be deleted.

Main.Sentinel1.GeolocationGrid — Type

GeolocationGrid

returns structure of GeolocationGrid from metadata in .xml

Main.Sentinel1.GeolocationGrid — Method

" GeolocationGrid

Constucture for the GeolocationGrid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the GeolocationGrid as a structure. Input in the GeolocationGrid file:
    lines: Reference image MDS line to which this geolocation grid point applies.
    samples,
    latitude: Geodetic latitude of grid point [degrees].
    longitude: Geodetic longitude of grid point [degrees].
    azimuthTime: Zero Doppler azimuth time to which grid point applies [UTC].
    slantRangeTime: Two-way slant range time to grid point [milli sec].
    elevationAngle: Elevation angle to grid point [degrees].
    incidenceAngle: Incidence angle to grid point [degrees].
    height: Height of the grid point above sea level [m].

Example:
    # accesing the geolocation data from the full metadata.
    xmlPath = "s1a-iw1-slc-vh-20220220t144146-20220220t144211-041998-050092-001.xml"
    Metadata1 = MetaDataSentinel1(xmlPath)
    geolocation = Metadata1.geolocation

    # accessing the geolocation directly from the xml.
    metaDict = getDictofXml(xmlPath)
    geolocation = GeolocationGrid(metaDict)
    

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
    GeolocationGrid[structure of GeolocationGrid]

Main.Sentinel1.Header — Type

Header

returns structure of Header from metadata in .xml

Main.Sentinel1.Header — Method

" Header

Constucture for the Header structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Header as a structure. Input in the header file:
    missionId: Mission identifier for this data set.
    productType: Product type for this data set.
    polarisation: Polarisation for this data set.
    missionDataTakeId: Mission data take identifier.
    swath: Swath identifier for this data set. This element identifies the swath that applies to all data contained within this data set. The swath identifier "EW" is used for products in which the 5 EW swaths have been merged. Likewise, "IW" is used for products in which the 3 IW swaths have been merged.
    mode: Sensor mode for this data set.
    startTime: Zero Doppler start time of the output image [UTC].
    stopTime: Zero Doppler stop time of the output image [UTC].
    aqusitionTime,
    absoluteOrbitNumber: Absolute orbit number at data set start time.
    imageNumber: Image number. For WV products the image number is used to distinguish between vignettes. For SM, IW and EW modes the image number is still used but refers instead to each swath and polarisation combination (known as the 'channel') of the data.

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
    Header[structure of Header]

Main.Sentinel1.ImageInformation — Type

ImageInformation

returns structure of ImageInformation from metadata in .xml

Main.Sentinel1.ImageInformation — Method

" ImageInformation

Constucture for the ImageInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the ImageInformation as a structure. Input in the ImageInformation file:
    rangePixelSpacing: Pixel spacing between range samples [m].
    azimuthFrequency: Azimuth line frequency of the output image [Hz]. This is the inverse of the azimuthTimeInterval.
    slantRangeTime: Two-way slant range time to first sample [milli sec].
    incidenceAngleMidSwath: Incidence angle at mid swath [degrees].
    azimuthPixelSpacing: Nominal pixel spacing between range lines [m].
    numberOfSamples: Total number of samples in the output image (image width).

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
    ImageInformation[structure of ImageInformation]

Main.Sentinel1.MetaDataSLC — Type

SwathSLC

A datatype for a Sentinel 1 Single Look Complex (SLC) swath subset.

Arguments

metadata::MetaDataSLC: Meta data for the Sentinel 1 swath

indexOffset::Tuple{Int,Int}: Pixel offeset of the subset with respect to the complete swath

pixels::Array{Complex,2}: The pixel values of the swath subset

Main.Sentinel1.MetaDataSentinel1 — Type

MetaDataSentinel1: Metadata structure for the Sentinel-1 satellite for each burst in the swath.

General metadata info is kept in the following structures:
    - Header
    - ProductInformation
    - ImageInformation
    - SwathTiming
    - GeolocationGrid
Burst specific Info is kept in 
    - BurstsInfo
Where BurstsInfo is a structure:
    numberOfBurst::Int64
    bursts::Vector{Burst}

Example: slcMetadata = MetaDataSentinel1(metaDict)

Input:
    metaDict: xml file.

can be accesed as, e.g., 
slcMetadata.product.radarFrequency --> 5.40500045433435e9::Float64
slcMetadata.header.swath --> 1::Int
slcMetadata.header.mode --> "IW"::String
slcMetadata.header.polarisation --> "VH"::String

Main.Sentinel1.MetaDataSentinel1 — Method

" MetaDataSentinel1

Constucture for the MetaDataSentinel1 structure. 
It takes a Sentinel-1 single swath metafile in .xml format and constructs the metadata structure using the individual sub-structures in the metadata.
The individual sub-structures in the metadata are:
- Header
- ProductInformation
- ImageInformation
- SwathTiming
- Burst
- GeolocationGrid

Input:
    xmlFile[string]: path of swath specific metadata in xml.format.

output:
    MetaDataSentinel1[structure of MetaDataSentinel1]: Strucutre with all Sentinel-1 metadata for a swath.

Example:

Getting the azimuthFmRateT0 for the 5th burst:
    metadata = MetaDataSentinel1(annotation.xml)
    metadata.burstsInfo.azimuthFmRate[5].azimuthFmRateT0

Main.Sentinel1.ProductInformation — Type

ProductInformation

returns structure of product information

Main.Sentinel1.ProductInformation — Method

" ProductInformation

Constucture for the ProductInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the ProductInformation as a structure. ProductInformation file:
    pass: Direction of the orbit (ascending, descending)
    timelinessCategory: Timeliness category under which the product was produced, i.e. time frame from the data acquisition
    platformHeading: Platform heading relative to North [degrees].
    projection: Projection of the image, either slant range or ground range.
    rangeSamplingRate: Range sample rate [Hz]
    radarFrequency: Radar (carrier) frequency [Hz]
    azimuthSteeringRate: Azimuth steering rate for IW and EW modes [degrees/s].

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
    ProductInformation[structure of ProductInformation]

Main.Sentinel1.SwathSLC — Type

SwathSLC

A datatype for a Sentinel 1 Single Look Complex (SLC) swath subset. 

# Arguments
- `metadata::MetaDataSLC`: Meta data for the Sentinel 1 swath
- `indexOffset::Tuple{Int,Int}`: Pixel offeset of the subset with respect to the complete swath
- `pixels::Array{Complex,2}`: The pixel values of the swath subset

Main.Sentinel1.SwathTiming — Type

SwathTiming

returns structure of SwathTiming from metadata in .xml

Main.Sentinel1.SwathTiming — Method

" SwathTiming

Constucture for the SwathTiming structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the SwathTiming as a structure. 

Input:
    metaDict[dict]: a dictionary of the metadata.

output:
    SwathTiming[structure of SwathTiming]

Main.Sentinel1.getAnnotationPaths — Method

getAnnotationPaths(safePath::string)

Getting the paths for the annotation files for a SLC image using its .SAFE folder path.

Parameters

* safePath::String: path of .SAFE folder for one image.

Returns

* annotationPaths::Vector: Vector of paths for annotation files in .SAFE folder

Main.Sentinel1.readMetaDataSLC — Method

readMetaDataSLC(filepath::String) Returns: ::MetaDataSLC

Main.Sentinel1.readSLC — Method

readSLC(folder::String, window) Returns: ::Array{SwathSLC,1}

Main.Sentinel1.readTiff — Function

readTiff(filepath::String, window=nothing, convertToDouble = true)

Read a Sentinel 1 tiff file.
# Examples:
```jldoctest
julia> filepath = "s1a-iw3-slc-vv-20220918t074921-20220918t074946-045056-056232-006.tiff"
julia> data = readSwathSLC(filePath, [(501,600),(501,650)]);
julia> typeof(data)
Matrix{ComplexF64}
julia> size(data)
(100,150)
```

Main.Sentinel1.searchDir — Method

"

search dir

Searching a directory for files with extention.