SARProcessing

Documentation for SARProcessing.

SARProcessing.Sentinel1AzimuthFmRate

SARProcessing.Sentinel1AzimuthFmRate

SARProcessing.Sentinel1BurstInformation

SARProcessing.Sentinel1BurstInformation

SARProcessing.Sentinel1DopplerCentroid

SARProcessing.Sentinel1DopplerCentroid

SARProcessing.Sentinel1GeolocationGrid

SARProcessing.Sentinel1GeolocationGrid

SARProcessing.Sentinel1Header

SARProcessing.Sentinel1Header

SARProcessing.Sentinel1ImageInformation

SARProcessing.Sentinel1ImageInformation

SARProcessing.Sentinel1MetaData

SARProcessing.Sentinel1MetaData

SARProcessing.Sentinel1ProductInformation

SARProcessing.Sentinel1ProductInformation

SARProcessing.Sentinel1SwathTiming

SARProcessing.Sentinel1SwathTiming

SARProcessing.TandemxDEM

SARProcessing._cell_averaging_constant_false_alarm_rate_pixel

SARProcessing.add_padding

SARProcessing.approx_line_of_sight

SARProcessing.azimuth_time2row

SARProcessing.binarize_array

SARProcessing.cell_averaging_constant_false_alarm_rate

SARProcessing.column2range

SARProcessing.constant_false_alarm_rate_with_convolution_and_pooling

SARProcessing.conv2d

SARProcessing.ecef2SAR_index

SARProcessing.ecef2geodetic

SARProcessing.ellipsoid_intersect

SARProcessing.geodetic2SAR_index

SARProcessing.geodetic2ecef

SARProcessing.get_burst_mid_times

SARProcessing.get_burst_numbers

SARProcessing.get_coordinate

SARProcessing.get_image_rows

SARProcessing.get_index

SARProcessing.get_sentinel1_annotation_paths

SARProcessing.get_subset

SARProcessing.get_window

SARProcessing.load_precise_orbit_sentinel1

SARProcessing.load_tandemx_dem

SARProcessing.load_tiff

SARProcessing.mask_array_nan

SARProcessing.mean_filter

SARProcessing.object_locations

SARProcessing.orbit_state_interpolator

SARProcessing.range2column

SARProcessing.row2azimuth_time

SARProcessing.sar2gray

SARProcessing.sar_index2geodetic

SARProcessing.search_directory

SARProcessing.sobel_filter

SARProcessing.solve_radar

SARProcessing.Sentinel1AzimuthFmRate — Type

Sentinel1AzimuthFmRate

returns structure of Sentinel1AzimuthFmRate from metadata in .xml Sentinel1AzimuthFmRate is calculated for each burst, and is therefore saved in each burst

SARProcessing.Sentinel1AzimuthFmRate — Method

" Sentinel1AzimuthFmRate

Constructors for the Sentinel1AzimuthFmRate structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1AzimuthFmRate data for a single burst as a structure. 

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., polynomial is redudant in the later processing. Could then be deleted.

SARProcessing.Sentinel1BurstInformation — Type

Sentinel1BurstInformation

returns structure of Sentinel1BurstInformation from metadata in .xml Sentinel1BurstInformation contain information from Sentinel1DopplerCentroid and Sentinel1AzimuthFmRate

SARProcessing.Sentinel1BurstInformation — Method

" Sentinel1BurstInformation

Constructors for the Sentinel1BurstInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1BurstInformation specific data for a single burst as a structure. 

Input:
    meta_dict[dict]: a dictionary of the metadata.burst_number

[Int]: Integer value of burst number.

output:
    Sentinel1BurstInformation[structure of Sentinel1BurstInformation]

SARProcessing.Sentinel1DopplerCentroid — Type

Sentinel1DopplerCentroid

returns structure of Sentinel1DopplerCentroid from metadata in .xml Sentinel1DopplerCentroid is calculated for each burst, and is therefore saved in each burst

SARProcessing.Sentinel1DopplerCentroid — Method

" Sentinel1DopplerCentroid

Constructors for the Sentinel1DopplerCentroid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1DopplerCentroid data for a single burst as a structure. 

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., polynomial is redudant in the later processing. Could be deleted.

SARProcessing.Sentinel1GeolocationGrid — Type

Sentinel1GeolocationGrid

returns structure of Sentinel1GeolocationGrid from metadata in .xml

SARProcessing.Sentinel1GeolocationGrid — Method

" Sentinel1GeolocationGrid

Constructors for the Sentinel1GeolocationGrid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1GeolocationGrid as a structure. Input in the Sentinel1GeolocationGrid file:
    lines: Reference image MDS line to which this geolocation grid point applies.
    samples,
    latitude: Geodetic latitude of grid point [degrees].
    longitude: Geodetic longitude of grid point [degrees].
    azimuth_time: Zero Doppler azimuth time to which grid point applies [UTC].
    slant_range_time_seconds: Two-way slant range time to grid point.
    elevation_angle: Elevation angle to grid point [degrees].
    incidence_angle: Incidence angle to grid point [degrees].
    height: Height of the grid point above sea level [m].

Example:
    # accesing the geolocation data from the full metadata.
    xmlPath = "s1a-iw1-slc-vh-20220220t144146-20220220t144211-041998-050092-001.xml"
    Metadata1 = Sentinel1MetaData(xmlPath)
    geolocation = Metadata1.geolocation

    # accessing the geolocation directly from the xml.
    meta_dict = read_xml_as_dict(xmlPath)
    geolocation = Sentinel1GeolocationGrid(meta_dict)
    

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1GeolocationGrid[structure of Sentinel1GeolocationGrid]

SARProcessing.Sentinel1Header — Type

Sentinel1Header

returns structure of Sentinel1Header from metadata in .xml

SARProcessing.Sentinel1Header — Method

" Sentinel1Header

Constructors for the Sentinel1Header structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1Header as a structure. Input in the header file:
    missionId: Mission identifier for this data set.
    productType: Product type for this data set.
    polarisation: Polarisation for this data set.
    mission_data_take_id: Mission data take identifier.
    swath: Swath identifier for this data set. This element identifies the swath that applies to all data contained within this data set. The swath identifier "EW" is used for products in which the 5 EW swaths have been merged. Likewise, "IW" is used for products in which the 3 IW swaths have been merged.
    mode: Sensor mode for this data set.
    start_time: Zero Doppler start time of the output image [UTC].
    stop_time: Zero Doppler stop time of the output image [UTC].
    absolute_orbit_number: Absolute orbit number at data set start time.
    image_number: Image number. For WV products the image number is used to distinguish between vignettes. For SM, IW and EW modes the image number is still used but refers instead to each swath and polarisation combination (known as the 'channel') of the data.

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1Header[structure of Sentinel1Header]

SARProcessing.Sentinel1ImageInformation — Type

Sentinel1ImageInformation

returns structure of Sentinel1ImageInformation from metadata in .xml

SARProcessing.Sentinel1ImageInformation — Method

" Sentinel1ImageInformation

Constructor for the Sentinel1ImageInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1ImageInformation as a structure. Input in the Sentinel1ImageInformation file:
    range_pixel_spacing: Pixel spacing between range samples [m].
    azimuth_frequency: Azimuth line frequency of the output image [Hz]. This is the inverse of the azimuth_timeInterval.
    slant_range_time_seconds: Two-way slant range time to first sample.
    incidence_angle_mid_swath: Incidence angle at mid swath [degrees].
    azimuth_pixel_spacing: Nominal pixel spacing between range lines [m].
    number_of_samples: Total number of samples in the output image (image width).

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1ImageInformation[structure of Sentinel1ImageInformation]

SARProcessing.Sentinel1MetaData — Type

Sentinel1MetaData: Metadata structure for the Sentinel-1 satellite for each burst in the swath.

General metadata info is kept in the following structures:
    - Sentinel1Header
    - Sentinel1ProductInformation
    - Sentinel1ImageInformation
    - Sentinel1SwathTiming
    - Sentinel1GeolocationGrid
Sentinel1BurstInformation specific Info is kept in 
    - Vector{Sentinel1BurstInformation}

Example: slcMetadata = Sentinel1MetaData(meta_dict)

Input:
    meta_dict: xml file.

can be accessed as, e.g., 
slcMetadata.product.radar_frequency --> 5.40500045433435e9::Float64
slcMetadata.header.swath --> 1::Int
slcMetadata.header.mode --> "IW"::String
slcMetadata.header.polarisation --> "VH"::String

SARProcessing.Sentinel1MetaData — Method

" Sentinel1MetaData

Constructors for the Sentinel1MetaData structure. 
It takes a Sentinel-1 single swath metafile in .xml format and constructs the metadata structure using the individual sub-structures in the metadata.
The individual sub-structures in the metadata are:
- Sentinel1Header
- Sentinel1ProductInformation
- Sentinel1ImageInformation
- Sentinel1SwathTiming
- Sentinel1BurstInformation
- Sentinel1GeolocationGrid

Input:
    xmlFile[string]: path of swath specific metadata in xml.format.

output:
    Sentinel1MetaData[structure of Sentinel1MetaData]: Structure with all Sentinel-1 metadata for a swath.

Example:

Getting the t0 for the 5th burst:
    metadata = Sentinel1MetaData(annotation.xml)
    metadata.bursts.azimuthFmRate[5].t0

SARProcessing.Sentinel1ProductInformation — Type

Sentinel1ProductInformation

returns structure of product information

SARProcessing.Sentinel1ProductInformation — Method

" Sentinel1ProductInformation

Constructors for the Sentinel1ProductInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1ProductInformation as a structure. Sentinel1ProductInformation file:
    pass: Direction of the orbit (ascending, descending)
    timeliness_category: Timeliness category under which the product was produced, i.e. time frame from the data acquisition
    platform_heading: Platform heading relative to North [degrees].
    projection: Projection of the image, either slant range or ground range.
    range_sampling_rate: Range sample rate [Hz]
    radar_frequency: Radar (carrier) frequency [Hz]
    azimuth_steering_rate: Azimuth steering rate for IW and EW modes [degrees/s].

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1ProductInformation[structure of Sentinel1ProductInformation]

SARProcessing.Sentinel1SwathTiming — Type

Sentinel1SwathTiming

returns structure of Sentinel1SwathTiming from metadata in .xml

SARProcessing.Sentinel1SwathTiming — Method

" Sentinel1SwathTiming

Constructors for the Sentinel1SwathTiming structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1SwathTiming as a structure. 

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1SwathTiming[structure of Sentinel1SwathTiming]

SARProcessing.TandemxDEM — Type

TandemxDEM <: DEM

A DEM implementation for Tandem-X DEM data.

SARProcessing._cell_averaging_constant_false_alarm_rate_pixel — Method

SARProcessing.add_padding — Function

Add padding to array

Q for simon:

SARProcessing.approx_line_of_sight — Method

approxlineofsight(orbitstate::OrbitState,incidenceanglemid::Real) # Output - line_of_sight::Array{float}(3): Line of sight to mid swath

#TODO, interpolate geolocationGridPoint from metadata instead?

SARProcessing.azimuth_time2row — Method

azimuthtime2row(azimuthtime::Real,metadata::MetaData)

Returns the row corresponding to a specific azimuth time.

Note: That the burst overlap is not considered in this function. 
The actual image row will thus differ.

SARProcessing.binarize_array — Method

binarize_array(image::Matrix{T}, threshold::Real= 0.0001) where T<:Real
Making a mask with boolean values of image

Examples:

    bool_array = binarize_array(rand(10,10), threshold= 0.0001)

SARProcessing.cell_averaging_constant_false_alarm_rate — Method

The CFAR object detection method described in The State-of-the-Art in Ship Detection in Synthetic Aperture Radar imagery, D.J. Crips, 2004, in section 5.2 Adaptive threshold algorithms

Finding CFAR for all pixels in an image.

SARProcessing.column2range — Method

column2range(column::Real,metadata::MetaData)

Returns the range corresponding to the image column

SARProcessing.constant_false_alarm_rate_with_convolution_and_pooling — Method

" The The constant false alarm rate with convolution and pooling (CP-CFAR) object detection method described in: Z. Cui, H. Quan, Z. Cao, S. Xu, C. Ding and J. Wu, "SAR Target CFAR Detection Via GPU Parallel Operation," in IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing, vol. 11, no. 12, pp. 4884-4894, Dec. 2018, doi: 10.1109/JSTARS.2018.2879082.

SARProcessing.conv2d — Function

Convolution function copied from Yosi Pramajaya. Credits goes to him. In his blogpost, he showed this implementation was faster than many others..
See https://towardsdatascience.com/understanding-convolution-by-implementing-in-julia-3ed744e2e933

dont want to have too many packages. I therefore wont use Convolution pkg.

Input:
    input::Matrix{Float64}. The input image,
    filter::Matrix{Float64}. The  filter/kernel to convolve
    stride::Int64 = 1. Stride of the convolution.
    padding::String = "valid". If padding is used ["valid" or "same"]

Output:
    result::Matrix{Float64}. convolved image.

# Example:
    #define a filter.
    average_pool_filter = filters.meanFilter([2,2])
    #perform convolution.
    image = operations.conv2d(image, average_pool_filter,2, "same")

SARProcessing.ecef2SAR_index — Method

ecef2SARindex( ecefcoordinate::Array{T,1}, interpolator, rangepixelspacing::Real, azimuthfrequency::Real, nearrange::Real, imagedurationseconds::Real ) where T <: Real

Convert ECEF-coordinates [X,Y,Z] to SARindex (rowfromfirstburst, image_column)

SARProcessing.ecef2geodetic — Method

ecef2geodetic(ecefcoordinate::Array{Real,1}; semimajoraxis=6378137., flattening=1/298.257223563, tolerancelatitude = 1.e-12, tolerance_height = 1.e-5)

Convert ECEF-coordinates [X,Y,Z] to geodetic-coordinates [latitude(radians),longitude(radians),height] (WGS-84) radians

(Based on B.R. Bowring, "The accuracy of geodetic latitude and height equations",
Survey Review, v28 #218, October 1985 pp.202-206).

SARProcessing.ellipsoid_intersect — Method

ellipsoidintersect(xsat::Array{Real,1},normalisedlineofsight::Array{Real,1}; semimajor_axis::Real=6378137.,flattening::Real=1/298.257223563)

Computes the intersection between the satellite line of sight and the earth ellipsoid in ECEF-coordinates
# Arguments
- `x_sat::Array{Real,1}`: [X,Y,Z] position of the satellite in ECEF-coordinates.
- `normalised_line_of_sight::Array{Real,1}`: Normalised Line of sight
# Output
- `x_0::Array{Real,1}`: intersection between line and ellipsoid in ECEF-coordinates.
# Note:
Equations follows I. Cumming and F. Wong (2005) p. 558-559

SARProcessing.geodetic2SAR_index — Method

geodetic2SARindex(geodeticcoordinate::Array{T,1}, interpolator, metadata::MetaData) where T <: Real

Convert geodetic-coordinates [latitude(radians),longitude(radians),height] to SARindex (rowfromfirstburst, image_column)

SARProcessing.geodetic2ecef — Method

geodetic2ecef(geodeticcoordinate::Array{Real,1}; semimajoraxis::Real=WGS84SEMIMAJORAXIS, flattening::Real=WGS84_FLATTENING)

Convert geodetic-coordinates [latitude(radians),longitude(radians),height] (WGS-84) to ECEF-coordinates [X,Y,Z]

SARProcessing.get_burst_mid_times — Method

getburstmid_times(image::T) where T <: SingleLookComplex

Returns a vector of the mid burst times for the burst in the image. 
Only bursts included in the image view are included

SARProcessing.get_burst_numbers — Method

getburstnumbers(image::Sentinel1SLC)

Returns list of burst included in the image subset

SARProcessing.get_coordinate — Method

get_coordinate(dem::T,index::Tuple{Integer,Integer}) where T <: DEM

Get the coordinate for a certain index in the DEM. 
(All the DEM's are stored as a matrix of heights)

SARProcessing.get_image_rows — Method

getimagerows(metadata::Sentinel1MetaData, rowfromfirstburst)

Converts the row number representing a unique azimuth time, rowfromfirst_burst, to the number in the full image. Note that two results are returned when the row appears in two bursts

SARProcessing.get_index — Method

get_index(dem::T,coordinate::Tuple{Real,Real}) where T <: DEM

Get the index of the DEM corresponding to the coordinate. 
(All the DEM's are stored as a matrix of heights)

SARProcessing.get_sentinel1_annotation_paths — Method

getsentinel1annotationpaths(safepath::string)

Getting the paths for the annotation files for a SLC image using its .SAFE folder path.

Parameters

* safe_path::String: path of .SAFE folder for one image.

Returns

* annotationPaths::Vector: Vector of paths for annotation files in .SAFE folder

SARProcessing.get_subset — Method

Extracting a subset from an image. The subset will be extraxted from the image row/column defined by coordinate and size subset_size.

Input:
    image: The image array
    coordinate::Vector{Int64}. Center coordinate of subset, in image geometry.
    subset_size::Vector{Int64}=[75,75]. Size of the subset. 

Output:
    subset::Array{Float64, 3}. The three dimensional subset [rows,columns,dimensions.] with dimension=1 for an input 2D array.

SARProcessing.get_window — Method

get_window(image::Sentinel1SLC)

Returns the window of the complete Sentinel image covered the "image"

SARProcessing.load_precise_orbit_sentinel1 — Method

loadpreciseorbit_sentinel1(path)

Loads a Sentinel 1 orbit file

Returns

Vector{OrbitState}

SARProcessing.load_tandemx_dem — Method

load_tandemx_dem(tiffPath::String) -> TandemxDEM

Load a Tandem-X DEM tiff. The Tandem-X DEM's can be downloaded from https://download.geoservice.dlr.de/TDM90/.
Note: Invalid values are replaced with missing

SARProcessing.load_tiff — Function

load_tiff(filepath::String, window=nothing; convertToDouble = true,flip = true)

Read a Sentinel 1 tiff file.
# Examples:
```jldoctest
julia> filepath = "s1a-iw3-slc-vv-20220918t074921-20220918t074946-045056-056232-006.tiff"
julia> data = readSwathSLC(filePath, [(501,600),(501,650)]);
julia> typeof(data)
Matrix{ComplexF64}
julia> size(data)
(100,150)
```

SARProcessing.mask_array_nan — Method

Making a mask with 1/NaN values of image

Examples:

    array_with_nan = mask_array_nan(rand(10,10), 0.5)

SARProcessing.mean_filter — Function

" meanFilter Creates a mean filter for a iamge

examples: filter = meanFilter([3,3]) or filter = meanFilter([3])

SARProcessing.object_locations — Method

"

Find center locations of objects in a binary image.

# Example:
    object_centers = object_locations(binary_array)

SARProcessing.orbit_state_interpolator — Method

orbitstateinterpolator(orbitstates::Vector{OrbitState}, image::SARImage, polynomialdegree::Integer=4, margin::Integer = 3 )

Create a polynomial interpolation function for orbit states valid in the time span
from image start time to image end time.

#Returns
Anonymous interpolation function. (Input: seconds_from_t_start::Float64, Output: state::OrbitState)

SARProcessing.range2column — Method

range2column(range::Real,metadata::MetaData)

Returns the image column corresponding to the range

SARProcessing.row2azimuth_time — Method

row2azimuthtime(rowfromfirstburst::Real,metadata::MetaData)

Returns the azimuth_time corresponding to a specific row (as counted from first burst ignoring burst overlap)

Note: That the burst overlap is not considered in this function.

SARProcessing.sar2gray — Method

sar2gray(data::AbstractArray; p_quantile = 0.85)

Maps the data to values between 0 and 1 and convert into a gray scaled image. The minimum data value is mapped to 0 and all values above the p_quantile is mapped to 1

SARProcessing.sar_index2geodetic — Method

sarindex2geodetic(rowfromfirstburst, image_column, height, interpolator, metadata::MetaData)

Convert SAR_index (row_from_first_burst, image_column) to geodetic coordinates [latitude(radians),longitude(radians),height]

SARProcessing.search_directory — Method

"

search dir

Searching a directory for files with extension.

SARProcessing.sobel_filter — Function

" sobelFilter(input::Matrix{Float64},stride::Int64 = 1,padding::String="same")::Matrix{Float64} Creates a sobelFilter for a image using edgeVertical() and edgeHorizontal()

examples: sobel_image = filters.sobelFilter(image)

SARProcessing.solve_radar — Method

solve_radar
Find the point that is range away from the satellite, orthogonal on the flight directions
and "height" above the elipsiod using Newton_rhapsody method.