SARProcessing

Documentation for SARProcessing.

SARProcessing.Sentinel1AzimuthFmRate

SARProcessing.Sentinel1AzimuthFmRate

SARProcessing.Sentinel1BurstInformation

SARProcessing.Sentinel1BurstInformation

SARProcessing.Sentinel1DopplerCentroid

SARProcessing.Sentinel1DopplerCentroid

SARProcessing.Sentinel1GeolocationGrid

SARProcessing.Sentinel1GeolocationGrid

SARProcessing.Sentinel1Header

SARProcessing.Sentinel1Header

SARProcessing.Sentinel1ImageInformation

SARProcessing.Sentinel1ImageInformation

SARProcessing.Sentinel1MetaData

SARProcessing.Sentinel1MetaData

SARProcessing.Sentinel1ProductInformation

SARProcessing.Sentinel1ProductInformation

SARProcessing.Sentinel1SwathTiming

SARProcessing.Sentinel1SwathTiming

SARProcessing.ecef2geodetic

SARProcessing.ellipsoid_intersect

SARProcessing.geodetic2ecef

SARProcessing.get_burst_mid_times

SARProcessing.get_burst_numbers

SARProcessing.get_sentinel1_annotation_paths

SARProcessing.get_time_range

SARProcessing.get_window

SARProcessing.load_precise_orbit_sentinel1

SARProcessing.load_tiff

SARProcessing.orbit_state_interpolator

SARProcessing.sar2gray

SARProcessing.search_directory

SARProcessing.Sentinel1AzimuthFmRate — Type

Sentinel1AzimuthFmRate

returns structure of Sentinel1AzimuthFmRate from metadata in .xml Sentinel1AzimuthFmRate is calculated for each burst, and is therefore saved in each burst

SARProcessing.Sentinel1AzimuthFmRate — Method

" Sentinel1AzimuthFmRate

Constructors for the Sentinel1AzimuthFmRate structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1AzimuthFmRate data for a single burst as a structure. 

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., polynomial is redudant in the later processing. Could then be deleted.

SARProcessing.Sentinel1BurstInformation — Type

Sentinel1BurstInformation

returns structure of Sentinel1BurstInformation from metadata in .xml Sentinel1BurstInformation contain information from Sentinel1DopplerCentroid and Sentinel1AzimuthFmRate

SARProcessing.Sentinel1BurstInformation — Type

" Sentinel1BurstInformation

Constructors for the Sentinel1BurstInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1BurstInformation specific data for a single burst as a structure. 

Input:
    meta_dict[dict]: a dictionary of the metadata.burst_number

[Int]: Integer value of burst number.

output:
    Sentinel1BurstInformation[structure of Sentinel1BurstInformation]

SARProcessing.Sentinel1DopplerCentroid — Type

Sentinel1DopplerCentroid

returns structure of Sentinel1DopplerCentroid from metadata in .xml Sentinel1DopplerCentroid is calculated for each burst, and is therefore saved in each burst

SARProcessing.Sentinel1DopplerCentroid — Method

" Sentinel1DopplerCentroid

Constructors for the Sentinel1DopplerCentroid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1DopplerCentroid data for a single burst as a structure. 

Note:
    [ ] Perhaps change input vars from strutures to the specific values. -- Does the current implementaion use extra time?
    [ ] What is needed? Maybe, e.g., polynomial is redudant in the later processing. Could be deleted.

SARProcessing.Sentinel1GeolocationGrid — Type

Sentinel1GeolocationGrid

returns structure of Sentinel1GeolocationGrid from metadata in .xml

SARProcessing.Sentinel1GeolocationGrid — Method

" Sentinel1GeolocationGrid

Constructors for the Sentinel1GeolocationGrid structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1GeolocationGrid as a structure. Input in the Sentinel1GeolocationGrid file:
    lines: Reference image MDS line to which this geolocation grid point applies.
    samples,
    latitude: Geodetic latitude of grid point [degrees].
    longitude: Geodetic longitude of grid point [degrees].
    azimuth_time: Zero Doppler azimuth time to which grid point applies [UTC].
    slant_range_time: Two-way slant range time to grid point [milli sec].
    elevation_angle: Elevation angle to grid point [degrees].
    incidence_angle: Incidence angle to grid point [degrees].
    height: Height of the grid point above sea level [m].

Example:
    # accesing the geolocation data from the full metadata.
    xmlPath = "s1a-iw1-slc-vh-20220220t144146-20220220t144211-041998-050092-001.xml"
    Metadata1 = Sentinel1MetaData(xmlPath)
    geolocation = Metadata1.geolocation

    # accessing the geolocation directly from the xml.
    meta_dict = read_xml_as_dict(xmlPath)
    geolocation = Sentinel1GeolocationGrid(meta_dict)
    

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1GeolocationGrid[structure of Sentinel1GeolocationGrid]

SARProcessing.Sentinel1Header — Type

Sentinel1Header

returns structure of Sentinel1Header from metadata in .xml

SARProcessing.Sentinel1Header — Method

" Sentinel1Header

Constructors for the Sentinel1Header structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1Header as a structure. Input in the header file:
    missionId: Mission identifier for this data set.
    productType: Product type for this data set.
    polarisation: Polarisation for this data set.
    mission_data_take_id: Mission data take identifier.
    swath: Swath identifier for this data set. This element identifies the swath that applies to all data contained within this data set. The swath identifier "EW" is used for products in which the 5 EW swaths have been merged. Likewise, "IW" is used for products in which the 3 IW swaths have been merged.
    mode: Sensor mode for this data set.
    start_time: Zero Doppler start time of the output image [UTC].
    stop_time: Zero Doppler stop time of the output image [UTC].
    absolute_orbit_number: Absolute orbit number at data set start time.
    image_number: Image number. For WV products the image number is used to distinguish between vignettes. For SM, IW and EW modes the image number is still used but refers instead to each swath and polarisation combination (known as the 'channel') of the data.

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1Header[structure of Sentinel1Header]

SARProcessing.Sentinel1ImageInformation — Type

Sentinel1ImageInformation

returns structure of Sentinel1ImageInformation from metadata in .xml

SARProcessing.Sentinel1ImageInformation — Method

" Sentinel1ImageInformation

Constructor for the Sentinel1ImageInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1ImageInformation as a structure. Input in the Sentinel1ImageInformation file:
    range_pixel_spacing: Pixel spacing between range samples [m].
    azimuth_frequency: Azimuth line frequency of the output image [Hz]. This is the inverse of the azimuth_timeInterval.
    slant_range_time: Two-way slant range time to first sample [milli sec].
    incidence_angle_mid_swath: Incidence angle at mid swath [degrees].
    azimuth_pixel_spacing: Nominal pixel spacing between range lines [m].
    number_of_samples: Total number of samples in the output image (image width).

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1ImageInformation[structure of Sentinel1ImageInformation]

SARProcessing.Sentinel1MetaData — Type

Sentinel1MetaData: Metadata structure for the Sentinel-1 satellite for each burst in the swath.

General metadata info is kept in the following structures:
    - Sentinel1Header
    - Sentinel1ProductInformation
    - Sentinel1ImageInformation
    - Sentinel1SwathTiming
    - Sentinel1GeolocationGrid
Sentinel1BurstInformation specific Info is kept in 
    - Vector{Sentinel1BurstInformation}

Example: slcMetadata = Sentinel1MetaData(meta_dict)

Input:
    meta_dict: xml file.

can be accessed as, e.g., 
slcMetadata.product.radar_frequency --> 5.40500045433435e9::Float64
slcMetadata.header.swath --> 1::Int
slcMetadata.header.mode --> "IW"::String
slcMetadata.header.polarisation --> "VH"::String

SARProcessing.Sentinel1MetaData — Method

" Sentinel1MetaData

Constructors for the Sentinel1MetaData structure. 
It takes a Sentinel-1 single swath metafile in .xml format and constructs the metadata structure using the individual sub-structures in the metadata.
The individual sub-structures in the metadata are:
- Sentinel1Header
- Sentinel1ProductInformation
- Sentinel1ImageInformation
- Sentinel1SwathTiming
- Sentinel1BurstInformation
- Sentinel1GeolocationGrid

Input:
    xmlFile[string]: path of swath specific metadata in xml.format.

output:
    Sentinel1MetaData[structure of Sentinel1MetaData]: Structure with all Sentinel-1 metadata for a swath.

Example:

Getting the t0 for the 5th burst:
    metadata = Sentinel1MetaData(annotation.xml)
    metadata.bursts.azimuthFmRate[5].t0

SARProcessing.Sentinel1ProductInformation — Type

Sentinel1ProductInformation

returns structure of product information

SARProcessing.Sentinel1ProductInformation — Method

" Sentinel1ProductInformation

Constructors for the Sentinel1ProductInformation structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1ProductInformation as a structure. Sentinel1ProductInformation file:
    pass: Direction of the orbit (ascending, descending)
    timeliness_category: Timeliness category under which the product was produced, i.e. time frame from the data acquisition
    platform_heading: Platform heading relative to North [degrees].
    projection: Projection of the image, either slant range or ground range.
    range_sampling_rate: Range sample rate [Hz]
    radar_frequency: Radar (carrier) frequency [Hz]
    azimuth_steering_rate: Azimuth steering rate for IW and EW modes [degrees/s].

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1ProductInformation[structure of Sentinel1ProductInformation]

SARProcessing.Sentinel1SwathTiming — Type

Sentinel1SwathTiming

returns structure of Sentinel1SwathTiming from metadata in .xml

SARProcessing.Sentinel1SwathTiming — Method

" Sentinel1SwathTiming

Constructors for the Sentinel1SwathTiming structure. 

It takes a dictionary containing the full sentinel-1 swath metadata and extracts the Sentinel1SwathTiming as a structure. 

Input:
    meta_dict[dict]: a dictionary of the metadata.

output:
    Sentinel1SwathTiming[structure of Sentinel1SwathTiming]

SARProcessing.ecef2geodetic — Method

ecef2geodetic(ecefcoordinate::Array{Real,1}; semimajoraxis=6378137., flattening=1/298.257223563, tolerancelatitude = 1.e-12, tolerance_height = 1.e-5)

Convert ECEF-coordinates [X,Y,Z] to geodetic-coordinates [latitude(radians),longitude(radians),height] (WGS-84) radians

(Based on B.R. Bowring, "The accuracy of geodetic latitude and height equations",
Survey Review, v28 #218, October 1985 pp.202-206).

SARProcessing.ellipsoid_intersect — Method

ellipsoidintersect(xsat::Array{Real,1},normalisedlineofsight::Array{Real,1}; semimajor_axis::Real=6378137.,flattening::Real=1/298.257223563)

Computes the intersection between the satellite line of sight and the earth ellipsoid in ECEF-coordinates
# Arguments
- `x_sat::Array{Real,1}`: [X,Y,Z] position of the satellite in ECEF-coordinates.
- `normalised_line_of_sight::Array{Real,1}`: Normalised Line of sight
# Output
- `x_0::Array{Real,1}`: intersection between line and ellipsoid in ECEF-coordinates.
# Note:
Equations follows I. Cumming and F. Wong (2005) p. 558-559

SARProcessing.geodetic2ecef — Method

geodetic2ecef(geodeticcoordinate::Array{Real,1}; semimajoraxis::Real=WGS84SEMIMAJORAXIS, flattening::Real=WGS84_FLATTENING)

Convert geodetic-coordinates [latitude(radians),longitude(radians),height] (WGS-84) to ECEF-coordinates [X,Y,Z]

SARProcessing.get_burst_mid_times — Method

getburstmid_times(image::T) where T <: SingleLookComplex

Returns a vector of the mid burst times for the burst in the image. 
Only bursts included in the image view are included

SARProcessing.get_burst_numbers — Method

getburstnumbers(image::Sentinel1SLC)

Returns list of burst included in the image subset

SARProcessing.get_sentinel1_annotation_paths — Method

getsentinel1annotationpaths(safepath::string)

Getting the paths for the annotation files for a SLC image using its .SAFE folder path.

Parameters

* safe_path::String: path of .SAFE folder for one image.

Returns

* annotationPaths::Vector: Vector of paths for annotation files in .SAFE folder

SARProcessing.get_time_range — Method

gettimerange(image::T) where T <: SARImage

#Returns
(start_time::DateTime, stop_time::DateTime)

SARProcessing.get_window — Method

get_window(image::Sentinel1SLC)

Returns the window of the complete Sentinel image covered the "image"

SARProcessing.load_precise_orbit_sentinel1 — Method

loadpreciseorbit_sentinel1(path)

Loads a Sentinel 1 orbit file

Returns

Vector{OrbitState}

SARProcessing.load_tiff — Function

load_tiff(filepath::String, window=nothing; convertToDouble = true,flip = true)

Read a Sentinel 1 tiff file.
# Examples:
```jldoctest
julia> filepath = "s1a-iw3-slc-vv-20220918t074921-20220918t074946-045056-056232-006.tiff"
julia> data = readSwathSLC(filePath, [(501,600),(501,650)]);
julia> typeof(data)
Matrix{ComplexF64}
julia> size(data)
(100,150)
```

SARProcessing.orbit_state_interpolator — Function

orbitstateinterpolator(orbitstates::Vector{OrbitState}, image::SARImage, polynomialdegree::Integer=4, margin::Integer = 3 )

Create a polynomial interpolation function for orbit states valid in the time span
from image start time to image end time.

#Returns
Anonymous interpolation function. (Input: DateTime, Output: OrbitState)

SARProcessing.sar2gray — Method

sar2gray(data::AbstractArray; p_quantile = 0.85)

Maps the data to values between 0 and 1 and convert into a gray scaled image. The minimum data value is mapped to 0 and all values above the p_quantile is mapped to 1

SARProcessing.search_directory — Method

"

search dir

Searching a directory for files with extension.