Sentinel1

Main.Sentinel1.MetaDataSLC — Type

SwathSLC

A datatype for a Sentinel 1 Single Look Complex (SLC) swath subset.

Arguments

metadata::MetaDataSLC: Meta data for the Sentinel 1 swath

indexOffset::Tuple{Int,Int}: Pixel offeset of the subset with respect to the complete swath

pixels::Array{Complex,2}: The pixel values of the swath subset

Main.Sentinel1.SwathSLC — Type

SwathSLC

A datatype for a Sentinel 1 Single Look Complex (SLC) swath subset. 

# Arguments
- `metadata::MetaDataSLC`: Meta data for the Sentinel 1 swath
- `indexOffset::Tuple{Int,Int}`: Pixel offeset of the subset with respect to the complete swath
- `pixels::Array{Complex,2}`: The pixel values of the swath subset

Main.Sentinel1.readMetaDataSLC — Method

readMetaDataSLC(filepath::String) Returns: ::MetaDataSLC

Main.Sentinel1.readSLC — Method

readSLC(folder::String, window) Returns: ::Array{SwathSLC,1}

Main.Sentinel1.readTiff — Function

readTiff(filepath::String, window=nothing, convertToDouble = true)

Read a Sentinel 1 tiff file.
# Examples:
```jldoctest
julia> filepath = "s1a-iw3-slc-vv-20220918t074921-20220918t074946-045056-056232-006.tiff"
julia> data = readSwathSLC(filePath, [(501,600),(501,650)]);
julia> typeof(data)
Matrix{ComplexF64}
julia> size(data)
(100,150)
```